Fixed open issues for release!
- [x] Ctrl for rebuilding the UI needs to be disabled for your release builds.
`Removed!`

- [x] Strange behavior when switching while a popout is active: https://gfycat.com/fakeresponsibleanemonecrab
`Changed to close on Template switch & clicking anywhere but the control`

- [x] I'm not sure what right-click in the Gear tab is meant to do.  It appears to just rebuild the UI (I guess it's pressing CTRL in the process).
`Working as intended, it copies the stat/upgrades name to your clipboard. It is in your case indeed the problem with CTRL triggering the Rebuild. Resetting the settings should fix, as the rebuild key is set to none by default now.`

- [x] Is revenant properly supported now?  If not, I would recommend removing it (or graying it out with a tooltip saying it's not ready) to avoid the questions/confusion that is sure to follow.
`No its not yet fully supported. Disclaimer added for Rev Builds`
![image](https://user-images.githubusercontent.com/102288477/167574156-1a997aa7-ab61-4ca8-8ff3-d74cf8c4f57d.png)

- [x] Module looks beautiful, though. Folks are going to absolutely love this.
`Thanks a lot :D`

- [x] Exception thrown by the AsyncWIndowsClipboard service
The first time I tried right-clicking in the Gear tab, I received a .NET exception that didn't leave a trace but said Access was denied and Blish HUD crashed.  If you're interacting with the clipboard (preferably through the Async clipboard service util we expose), I'd recommend catching any errors if there is a chance they bubble up.
`Checking the clipboard content if its null / "" and wrapped in try/catch`